+++ tests/unit/CRUD/odata/odata-batch-reading-test.js
@@ -1,0 +1,6 @@
+import executeTest from './execute-odata-test';
+import batchReadingFunctions from '../base/base-batch-reading-test';
+
+executeTest('reading | batch reading', (store, assert) => {
+  batchReadingFunctions(store, assert);
+});