+++ tests/unit/CRUD/offline/offline-batch-reading-test.js
@@ -1,0 +1,6 @@
+import executeTest from './execute-offline-test';
+import batchReadingFunctions from '../base/base-batch-reading-test';
+
+executeTest('reading | batch reading', (store, assert) => {
+  batchReadingFunctions(store, assert);
+});